+++ package.json
@@ -10,6 +10,7 @@
     "bootstrap": "^5.3.2",
     "cors": "^2.8.5",
     "express": "^4.18.2",
+    "jwt-decode": "^4.0.0",
     "mongoose": "^8.0.1",
     "nodemon": "^3.0.1",
     "react": "^18.2.0",
