+++ src/component/auth/auth-helper.js
@@ -9,4 +9,33 @@ const authenticate = (token, cb)=>{
   }
   cb();
 }
-export {authenticate};
+
+const isAuthenticated = ()=>{
+  if (typeof window === "undefined") {
+    return false;
+  }
+  return !!sessionStorage.getItem('token');
+}
+
+const getToken = ()=>{
+  if (typeof window === "undefined") {
+    return false;
+  }
+  return sessionStorage.getItem('token');
+}
+
+const getUsername = ()=>{
+  if (typeof window === "undefined") {
+    return false;
+  }
+  return sessionStorage.getItem('username');
+}
+
+const clearJWT = ()=>{
+  if (typeof window !== "undefined") {
+    sessionStorage.removeItem('token');
+    sessionStorage.removeItem('username');
+  }
+}
+
+export { authenticate, isAuthenticated, getToken, getUsername, clearJWT }
