+++ src/pages/Signup.js
@@ -1,57 +1,57 @@
-import React,{ useEffect,useState} from "react"
-import axios from "axios"
-import {useNavigate, Link} from "react-router-dom"
+// import React,{ useEffect,useState} from "react"
+// import axios from "axios"
+// import {useNavigate, Link} from "react-router-dom"
 
-function Signup() {
-    const history=useNavigate();
-    const [email,setEmail]=useState('')
-    const [password,setPassword]=useState('')
+// function Signup() {
+//     const history=useNavigate();
+//     const [email,setEmail]=useState('')
+//     const [password,setPassword]=useState('')
 
-    async function submit(e) {
-        e.preventDefault();
+//     async function submit(e) {
+//         e.preventDefault();
 
-        try{
-            await axios.post("http://localhost:8000/signup",{
-                email,password
-            })
-            .then(res=>{
-                if(res.data=="exist"){
-                    alert("User already exists")
-                }
-                else if(res.data=="notexist"){
-                    history("/Home",{state:{id:email}})
-                }
-            })
-            .catch(e=>{
-                alert("wrong details")
-                console.log(e);
-            })
-        }
-        catch(e){
-            console.log(e);
-        }
-    }
+//         try{
+//             await axios.post("http://localhost:8000/signup",{
+//                 email,password
+//             })
+//             .then(res=>{
+//                 if(res.data=="exist"){
+//                     alert("User already exists")
+//                 }
+//                 else if(res.data=="notexist"){
+//                     history("/Home",{state:{id:email}})
+//                 }
+//             })
+//             .catch(e=>{
+//                 alert("wrong details")
+//                 console.log(e);
+//             })
+//         }
+//         catch(e){
+//             console.log(e);
+//         }
+//     }
 
-    return (
-        <div className="login">
+//     return (
+//         <div className="login">
 
-            <h1>Signup</h1>
+//             <h1>Signup</h1>
 
-            <form action="POST">
-                <input type="email" onChange={(e) => { setEmail(e.target.value) }} placeholder="Email"  />
-                <input type="password" onChange={(e) => { setPassword(e.target.value) }} placeholder="Password"  />
-                <input type="submit" onClick={submit} />
+//             <form action="POST">
+//                 <input type="email" onChange={(e) => { setEmail(e.target.value) }} placeholder="Email"  />
+//                 <input type="password" onChange={(e) => { setPassword(e.target.value) }} placeholder="Password"  />
+//                 <input type="submit" onClick={submit} />
 
-            </form>
+//             </form>
 
-            <br />
-            <p>OR</p>
-            <br />
+//             <br />
+//             <p>OR</p>
+//             <br />
 
-            <Link to="/Login">Login Page</Link>
+//             <Link to="/Login">Login Page</Link>
 
-        </div>
-    )
-}
+//         </div>
+//     )
+// }
 
-export default Signup
+// export default Signup
